src/requirements.txt
amqp==5.0.9
billiard==3.6.4.0
celery==5.2.3
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==4.0.2
djangorestframework==3.13.1
gunicorn==20.1.0
kombu==5.2.3
packaging==21.3
prompt-toolkit==3.0.28
pyparsing==3.0.7
pytz==2021.3
psycopg2-binary==2.9.3
redis==4.1.3
six==1.16.0
sqlparse==0.4.2
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
requests==2.27.1
urllib3==1.26.8
